Separate slice as module
